feat: added generic column definitions + telemetry table column definitions
