+++ app/src/features/surveys/observations/observations-table/grid-column-definitions/GridColumnDefinitions.tsx
@@ -10,12 +10,8 @@ import ConditionalAutocompleteDataGridViewCell from 'components/data-grid/condit
 import TaxonomyDataGridEditCell from 'components/data-grid/taxonomy/TaxonomyDataGridEditCell';
 import TaxonomyDataGridViewCell from 'components/data-grid/taxonomy/TaxonomyDataGridViewCell';
 import TextFieldDataGrid from 'components/data-grid/TextFieldDataGrid';
-import TimePickerDataGrid from 'components/data-grid/TimePickerDataGrid';
-import { DATE_FORMAT } from 'constants/dateTimeFormats';
 import { IObservationTableRow } from 'contexts/observationsTableContext';
-import { default as dayjs } from 'dayjs';
 import { CBMeasurementType, CBQualitativeOption } from 'interfaces/useCritterApi.interface';
-import { getFormattedDate } from 'utils/Utils';
 
 export type ISampleSiteOption = {
   survey_sample_site_id: number;
@@ -249,229 +245,6 @@ export const ObservationCountColDef = (props: {
             onChange: (event) => {
               if (!/^\d{0,7}$/.test(event.target.value)) {
                 // If the value is not a number, return
-                return;
-              }
-
-              params.api.setEditCellValue({
-                id: params.id,
-                field: params.field,
-                value: event.target.value
-              });
-            },
-            error
-          }}
-        />
-      );
-    }
-  };
-};
-
-export const ObservationDateColDef = (props: {
-  hasError: (params: GridCellParams) => boolean;
-}): GridColDef<IObservationTableRow> => {
-  const { hasError } = props;
-
-  return {
-    field: 'observation_date',
-    headerName: 'Date',
-    editable: true,
-    hideable: true,
-    type: 'date',
-    minWidth: 150,
-    valueGetter: (params) => (params.row.observation_date ? dayjs(params.row.observation_date).toDate() : null),
-    disableColumnMenu: true,
-    headerAlign: 'left',
-    align: 'left',
-    renderCell: (params) => (
-      <Typography variant="body2" sx={{ fontSize: 'inherit' }}>
-        {getFormattedDate(DATE_FORMAT.ShortDateFormatMonthFirst, params.value)}
-      </Typography>
-    ),
-    renderEditCell: (params) => {
-      const error = hasError(params);
-
-      return (
-        <TextFieldDataGrid
-          dataGridProps={params}
-          textFieldProps={{
-            name: params.field,
-            type: 'date',
-            value: params.value ? dayjs(params.value).format('YYYY-MM-DD') : '',
-            onChange: (event) => {
-              const value = dayjs(event.target.value).toDate();
-
-              params.api.setEditCellValue({
-                id: params.id,
-                field: params.field,
-                value
-              });
-            },
-
-            error
-          }}
-        />
-      );
-    }
-  };
-};
-
-export const ObservationTimeColDef = (props: {
-  hasError: (params: GridCellParams) => boolean;
-}): GridColDef<IObservationTableRow> => {
-  const { hasError } = props;
-
-  return {
-    field: 'observation_time',
-    headerName: 'Time',
-    editable: true,
-    hideable: true,
-    type: 'string',
-    width: 150,
-    disableColumnMenu: true,
-    headerAlign: 'right',
-    align: 'right',
-    valueSetter: (params) => {
-      return { ...params.row, observation_time: params.value };
-    },
-    valueParser: (value) => {
-      if (!value) {
-        return null;
-      }
-
-      if (dayjs.isDayjs(value)) {
-        return value.format('HH:mm:ss');
-      }
-
-      return dayjs(value, 'HH:mm:ss').format('HH:mm:ss');
-    },
-    renderCell: (params) => {
-      if (!params.value) {
-        return null;
-      }
-
-      return (
-        <Typography variant="body2" sx={{ fontSize: 'inherit' }}>
-          {params.value}
-        </Typography>
-      );
-    },
-    renderEditCell: (params) => {
-      const error = hasError(params);
-
-      return (
-        <TimePickerDataGrid
-          dataGridProps={params}
-          dateFieldProps={{
-            slotProps: {
-              textField: {
-                error,
-                name: params.field
-              }
-            }
-          }}
-        />
-      );
-    }
-  };
-};
-
-export const ObservationLatitudeColDef = (props: {
-  hasError: (params: GridCellParams) => boolean;
-}): GridColDef<IObservationTableRow> => {
-  const { hasError } = props;
-
-  return {
-    field: 'latitude',
-    headerName: 'Lat',
-    editable: true,
-    hideable: true,
-    width: 120,
-    disableColumnMenu: true,
-    headerAlign: 'right',
-    align: 'right',
-    valueSetter: (params) => {
-      if (/^-?\d{1,3}(?:\.\d{0,12})?$/.test(params.value)) {
-        // If the value is a legal latitude value
-        // Valid entries: `-1`, `-1.1`, `-123.456789` `1`, `1.1, `123.456789`
-        return { ...params.row, latitude: Number(params.value) };
-      }
-
-      const value = parseFloat(params.value);
-      return { ...params.row, latitude: isNaN(value) ? null : value };
-    },
-    renderCell: (params) => (
-      <Typography variant="body2" sx={{ fontSize: 'inherit' }}>
-        {params.value}
-      </Typography>
-    ),
-    renderEditCell: (params) => {
-      const error: boolean = hasError(params);
-
-      return (
-        <TextFieldDataGrid
-          dataGridProps={params}
-          textFieldProps={{
-            name: params.field,
-            onChange: (event) => {
-              if (!/^-?\d{0,3}(?:\.\d{0,12})?$/.test(event.target.value)) {
-                // If the value is not a subset of a legal latitude value, prevent the value from being applied
-                return;
-              }
-
-              params.api.setEditCellValue({
-                id: params.id,
-                field: params.field,
-                value: event.target.value
-              });
-            },
-            error
-          }}
-        />
-      );
-    }
-  };
-};
-
-export const ObservationLongitudeColDef = (props: {
-  hasError: (params: GridCellParams) => boolean;
-}): GridColDef<IObservationTableRow> => {
-  const { hasError } = props;
-
-  return {
-    field: 'longitude',
-    headerName: 'Long',
-    editable: true,
-    hideable: true,
-    width: 120,
-    disableColumnMenu: true,
-    headerAlign: 'right',
-    align: 'right',
-    valueSetter: (params) => {
-      if (/^-?\d{1,3}(?:\.\d{0,12})?$/.test(params.value)) {
-        // If the value is a legal longitude value
-        // Valid entries: `-1`, `-1.1`, `-123.456789` `1`, `1.1, `123.456789`
-        return { ...params.row, longitude: Number(params.value) };
-      }
-
-      const value = parseFloat(params.value);
-      return { ...params.row, longitude: isNaN(value) ? null : value };
-    },
-    renderCell: (params) => (
-      <Typography variant="body2" sx={{ fontSize: 'inherit' }}>
-        {params.value}
-      </Typography>
-    ),
-    renderEditCell: (params) => {
-      const error: boolean = hasError(params);
-
-      return (
-        <TextFieldDataGrid
-          dataGridProps={params}
-          textFieldProps={{
-            name: params.field,
-            onChange: (event) => {
-              if (!/^-?\d{0,3}(?:\.\d{0,12})?$/.test(event.target.value)) {
-                // If the value is not a subset of a legal longitude value, prevent the value from being applied
                 return;
               }
